requirements.txt
fastapi
uvicorn
python-multipart
pillow
google-generativeai
python-dotenv
google-genai